Major reorganisation of packages
Some classes are not grouped by concrete functionality.

Packages should group classes that interact heavily with each other together.

Moving classes according to the MVC architecture makes the overall structure much more prominent and easy to understand
