+++ src/main/java/uwuke/DukeException.java
@@ -1,4 +1,4 @@
-package uwuke.output;
+package uwuke;
 
 public class DukeException extends Exception {
     public DukeException(String e) {
