+++ src/main/java/uwuke/model/Event.java
@@ -1,4 +1,4 @@
-package uwuke.task;
+package uwuke.model;
 
 public class Event extends Task {
