+++ src/main/java/uwuke/model/Todo.java
@@ -1,4 +1,4 @@
-package uwuke.task;
+package uwuke.model;
 
 public class Todo extends Task {
