+++ src/main/java/uwuke/view/DialogBox.java
@@ -1,4 +1,4 @@
-package uwuke.gui;
+package uwuke.view;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
